+++ __init__.py
@@ -67,6 +67,8 @@ class EZB_preferences(bpy.types.AddonPreferences):
         max=59
     )
 
+    run_in_background: bpy.props.BoolProperty(default=True, name='Run Bakes in background', description='If active, blender will be responsive while the images are being generated')
+
     def set_abs_handplane_path(self, value):
         new_path = os.path.abspath(bpy.path.abspath(value))
         self.abs_handplane_path = new_path
@@ -77,9 +79,21 @@ class EZB_preferences(bpy.types.AddonPreferences):
     abs_handplane_path: StringProperty(default='C:\\Program Files\\Handplane3D LLC\\Handplane Baker\\')
     handplane_path: bpy.props.StringProperty(subtype='DIR_PATH', set=set_abs_handplane_path, get=get_abs_handplane_path)
 
+    def set_abs_marmoset_path(self, value):
+        new_path = os.path.abspath(bpy.path.abspath(value))
+        self.abs_marmoset_path = new_path
+
+    def get_abs_marmoset_path(self):
+        return self.abs_marmoset_path
+
+    abs_marmoset_path: StringProperty(default='C:\\Program Files\\Marmoset\\Toolbag 3\\')
+    marmoset_path: bpy.props.StringProperty(subtype='DIR_PATH', set=set_abs_marmoset_path, get=get_abs_marmoset_path)
+
     def draw(self, context):
         layout = self.layout
+        #layout.prop(self, 'run_in_background')
         layout.prop(self, 'handplane_path')
+        #layout.prop(self, 'marmoset_path')
         ops.update_settings_ui(self, context)